--- app/src/main/java/com/completewordproblems/fourthgrade/MyNavHostFragment.kt
@@ -1,7 +1,0 @@
-package com.completewordproblems.fourthgrade
-
-import androidx.navigation.fragment.NavHostFragment
-
-class MyNavHostFragment() : NavHostFragment() {
-
-}
