+++ app/src/main/java/com/completewordproblems/fourthgrade/fragments/MyNavigationFragment.kt
@@ -1,10 +1,13 @@
-package com.completewordproblems.fourthgrade
+package com.completewordproblems.fourthgrade.fragments
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Button
+import androidx.fragment.app.Fragment
+import androidx.navigation.findNavController
+import com.completewordproblems.fourthgrade.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +16,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [WhatAreYouLookingForFragment.newInstance] factory method to
+ * Use the [MyNavigationFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class WhatAreYouLookingForFragment : Fragment() {
+class MyNavigationFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -33,8 +36,19 @@ class WhatAreYouLookingForFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_what_are_you_looking_for, container, false)
+        val view: View = inflater.inflate(R.layout.fragment_my_navigation, container, false)
+        view.findViewById<Button>(R.id.practice_button).setOnClickListener {
+            view.findNavController().navigate(R.id.action_myNavigationFragment_to_practiceFragment)
+        }
+        view.findViewById<Button>(R.id.growth_button).setOnClickListener {
+            view.findNavController()
+                .navigate(R.id.action_myNavigationFragment_to_yourGrowthFragment)
+        }
+        view.findViewById<Button>(R.id.strategy_button).setOnClickListener {
+            view.findNavController()
+                .navigate(R.id.action_myNavigationFragment_to_strategyAlgorithmFragment)
+        }
+        return view
     }
 
     companion object {
@@ -44,12 +58,12 @@ class WhatAreYouLookingForFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment WhatAreYouLookingForFragment.
+         * @return A new instance of fragment MyNavigationFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            WhatAreYouLookingForFragment().apply {
+            MyNavigationFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
