+++ app/src/main/java/com/completewordproblems/fourthgrade/fragments/SolveFragment.kt
@@ -1,10 +1,11 @@
-package com.completewordproblems.fourthgrade
+package com.completewordproblems.fourthgrade.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.completewordproblems.fourthgrade.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [YourGrowthFragment.newInstance] factory method to
+ * Use the [SolveFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class YourGrowthFragment : Fragment() {
+class SolveFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +35,7 @@ class YourGrowthFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_your_growth, container, false)
+        return inflater.inflate(R.layout.fragment_solve, container, false)
     }
 
     companion object {
@@ -44,12 +45,12 @@ class YourGrowthFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment YourGrowthFragment.
+         * @return A new instance of fragment SolveFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            YourGrowthFragment().apply {
+            SolveFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
