+++ app/src/main/java/com/completewordproblems/fourthgrade/fragments/YourGrowthFragment.kt
@@ -1,10 +1,11 @@
-package com.completewordproblems.fourthgrade
+package com.completewordproblems.fourthgrade.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.completewordproblems.fourthgrade.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [WhatIsGivenFragment.newInstance] factory method to
+ * Use the [YourGrowthFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class WhatIsGivenFragment : Fragment() {
+class YourGrowthFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +35,7 @@ class WhatIsGivenFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_what_is_given, container, false)
+        return inflater.inflate(R.layout.fragment_your_growth, container, false)
     }
 
     companion object {
@@ -44,12 +45,12 @@ class WhatIsGivenFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment WhatIsGivenFragment.
+         * @return A new instance of fragment YourGrowthFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            WhatIsGivenFragment().apply {
+            YourGrowthFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
